data type

+++ demo.py
@@ -160,7 +160,7 @@ logging.getLogger("rsciio.emd").setLevel(logging.ERROR)
 
 
 save_path = projects_dir + "/tmp/unaligned_hsi"
-tmp = store_unaligned_hsi_alt(file_path, save_path, n_frames=num_frames, data_type='float16')
+tmp = store_unaligned_hsi_alt(file_path, save_path, n_frames=num_frames, data_type='float32')
 print("The unaligned HSI has been stored in: %s " % save_path)
